feat: add re-establish wallet modals
