+++ lib/screens/login/view/login_screen.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:my_wit_wallet/screens/login/view/biometrics_autentication.dart';
-import 'package:my_wit_wallet/screens/login/view/re_establish_wallet.dart';
+import 'package:my_wit_wallet/screens/login/view/re_establish_wallet_button.dart';
 import 'package:my_wit_wallet/theme/wallet_theme.dart';
 import 'package:my_wit_wallet/util/storage/database/wallet_storage.dart';
 import 'package:my_wit_wallet/widgets/layouts/layout.dart';
@@ -228,7 +228,7 @@ class LoginScreenState extends State<LoginScreen>
                             ? BiometricsAutentication()
                             : Container(),
                         SizedBox(height: 8),
-                        ReEstablishWallet(),
+                        ReEstablishWalletBtn(),
                       ])
                 ],
                 actions: [_loginListener()],
